pyproject.toml

[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "devin-zero-etl-demo"
version = "0.1.0"
description = "Salesforce → Oracle transformations (Zero ETL demo)"
authors = [{name="Demo"}]
dependencies = ["pandas", "python-dateutil"]
